chore: split engine feat: priority filtering feat: parallel enrich tasks
